BAU: Move Cognito client and lambda facade to services
