+++ express/src/services/cognito/CognitoClient.interface.d.ts
