+++ express/src/services/lambda/LambdaFacade.ts
@@ -1,6 +1,3 @@
-import Service from "@self-service/client/service";
-import User from "@self-service/client/user";
-import OnboardingTableItem from "@self-service/common/onboarding-table-item";
 import axios, {Axios, AxiosResponse} from "axios";
 import {OnboardingTableItem} from "../../../@types/OnboardingTableItem";
 import {Service} from "../../../@types/Service";
