+++ express/src/services/lambda/StubLambdaFacade.ts
